Made imageBuildTime Instant
